+++ lib/src/dialogs/dialogs.dart
@@ -1,6 +1,4 @@
 export 'actions.dart';
-export 'basic_alert_dialog.dart';
 export 'basic_bottom_sheet.dart';
 export 'platform_bottom_sheet.dart';
-export 'platform_dialog.dart';
 export 'platform_widget.dart';